Rename 'role' to 'name' in RFC 0018 and PoC plan
Align the docs with the code rename: a tensor's lookup key is a 'name', not a
'role'. Rewords the concept throughout (op-schema names, name-map/name-bind
grammar, A.3 'Names and edges', B.5 registry example keys) and lowercases the
q/k/v/o example strings to match the schema.

+++ projects/hipdnn/flatbuffers_sdk/tests/umd/TestOpSchemaRegistry.cpp
@@ -78,8 +78,7 @@ fb::FlatBufferBuilder buildSdpa()
 TEST(UmdOpSchemaRegistry, SdpaEntryResolvesByNameAndType)
 {
     const auto* byName = umd::lookupOpByName("sdpa_fwd"); // the umd_opcode shorthand
-    const auto* byType
-        = umd::lookupOpByType(static_cast<int>(data::NodeAttributes::SdpaAttributes));
+    const auto* byType = umd::lookupOpByType(data::NodeAttributes::SdpaAttributes);
     ASSERT_NE(byName, nullptr);
     ASSERT_NE(byType, nullptr);
     EXPECT_EQ(byName, byType);
